Add example playbook to create a FAI-like server
We need some kind of bare metal provisioning / automated installation
system. We tried MaaS but it is too restrictive on disk partitioning and
there is no Debian support. So we have this combination of DHCP,
di-netboot-assistant, DI preseed and a few scripts to create automated
installation both for PXE and UEFI.

This playbook serves as an example on how to set up the install server
in Ignum test IP space and all.

TODO:
Parametrize - namely the IP addresses, default passwords and names should
be placed in variables.
DNS and host naming support.
NetBox binding.